Add theme support and settings page to app
Two main enhancements have been made to the application: Firstly, support for both light and dark themes was introduced, with the themes being applied to the MaterialApp definition in app.dart file. Secondly, a new settings page was added to the app's navigation. Also, the go_router package was removed and app routes were updated accordingly.